pyproject.toml
[tool.poetry]
name = "circuit-cli"
version = "0.1.0"
description = ""
authors = ["Sebastjan Trepca <set@circuitdao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
chia-blockchain = "^2.4.1rc1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
circuit-cli = "circuit_cli.circuit_rpc_cli:main"
price-announcer-cli = "circuit_cli.price_announcer_cli:main"
keeper-cli = "circuit_cli.keeper_cli:main"
